File-Operations
When the program is executed, it should perform the search operation described in Application 2 within a folder provided as an argument, encompassing all files in that folder (excluding subfolders). However, the search should only be carried out in files in Portable Executable (PE) format. Other files should be skipped, and this information should be conveyed to the user. To determine whether a file is in PE format, it is sufficient to check whether the first 2 bytes are "MZ." For this example, there is no need to implement the actual validation processes. Once it is determined that the file is in PE format, a search should be conducted only within the boundaries of sections that have the "executable" flag. In other words, while searching for the "MALWARE" string, the search should only take place within sections that have the "executable" flag. If the string is found, the location along with the section name should be provided to the user."When the program is executed, it should perform a search operation within a folder provided as an argument, encompassing all files in that folder (excluding subfolders). However, the search should only be carried out in files in Portable Executable (PE) format. Other files should be skipped, and this information should be conveyed to the user. To determine whether a file is in PE format, it is sufficient to check whether the first 2 bytes are "MZ." For this example, there is no need to implement the actual validation processes. Once it is determined that the file is in PE format, a search should be conducted only within the boundaries of sections that have the "executable" flag. In other words, while searching for the "MALWARE" string, the search should only take place within sections that have the "executable" flag. If the string is found, the location along with the section name should be provided to the user."